fix(android): rename the root package from com.sample to com.react.testapp
