+++ android/app/src/main/java/com/react/testapp/manifest/Manifest.kt
@@ -1,4 +1,4 @@
-package com.sample.manifest
+package com.react.testapp.manifest
 
 import com.squareup.moshi.JsonClass
