+++ android/app/src/main/java/com/react/testapp/manifest/ManifestModule.kt
@@ -1,4 +1,4 @@
-package com.sample.manifest
+package com.react.testapp.manifest
 
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.Moshi
