+++ android/app/src/test/java/com/react/testapp/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.sample
+package com.react.testapp
 
 import org.junit.Test
